Cleaned up linear regression code.
Code is closer to check*py .
Added plotting code (6 subplots for a stock) to utils.py
Added a string with parameters to the plot chart outputs

Notes:
Linear regression is over n (often 200) days, using "n days worth of features,
such as "close" and "vol". Then, I do prediction over "m" subsequent days.
That is not necessarily the best way to use linear regression.

Perhaps I should apply linear regression with 1-2 days worth of features over
a much smaller number of days, say 10 or 20 and then predict only the next day.
So instead of setting up a single linear fit with 200 nodes, one sets up a linear
fit each time a new prediction is to be made.
This approach is more amenable to a neural network.